@@ -9,6 +9,7 @@ import (
 	"syscall"
 	"time"
 
+	"github.com/twbworld/proxy/dao"
 	"github.com/twbworld/proxy/global"
 	"github.com/twbworld/proxy/router"
 	"github.com/twbworld/proxy/service"
@@ -16,15 +17,16 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func Init() {
+var server *http.Server
+
+func init() {
 
 	ginfile, err := os.OpenFile(global.Config.AppConfig.GinLogPath, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
 	if err != nil {
 		panic("打开文件错误[nvcbjkdfgu]: " + err.Error())
 	}
-	gin.DefaultWriter = io.MultiWriter(ginfile) //记录所有日志
-	gin.DefaultErrorWriter = global.Log.Out
-	gin.DisableConsoleColor() //将日志写入文件时不需要控制台颜色
+	gin.DefaultWriter, gin.DefaultErrorWriter = io.MultiWriter(ginfile), global.Log.Out //记录所有日志
+	gin.DisableConsoleColor()                                                           //将日志写入文件时不需要控制台颜色
 	mode := gin.ReleaseMode
 	if global.Config.Env.Debug {
 		mode = gin.DebugMode
@@ -32,13 +34,19 @@ func Init() {
 	gin.SetMode(mode)
 
 	ginServer := gin.Default()
-	router.Init(ginServer)
+	router.Start(ginServer)
 
 	// ginServer.Run(":80")
-	server := &http.Server{
+	server = &http.Server{
 		Addr:    global.Config.AppConfig.GinAddr,
 		Handler: ginServer,
 	}
+
+}
+
+func Start() {
+	dao.Start()
+	TgStart()
 
 	//协程启动服务
 	go func() {
@@ -46,8 +54,6 @@ func Init() {
 			global.Log.Panic("服务出错[isjfio]: ", err.Error()) //外部并不能捕获Panic
 		}
 	}()
-
-	TgInit()
 
 	global.Log.Infof("启动成功, port: %s, pid: %d", global.Config.AppConfig.GinAddr, syscall.Getpid())
 	service.TgSend("启动成功")
