@@ -5,7 +5,7 @@ import (
 	"github.com/twbworld/proxy/global"
 )
 
-func TgInit() {
+func TgStart() {
 	if global.Config.Env.Debug || global.Config.Env.Telegram.Token == "" {
 		return
 	}
@@ -25,26 +25,27 @@ func TgInit() {
 	if _, err := global.Bot.Request(setCommands); err != nil {
 		global.Log.Warnln("设置Command失败[podritgfd]: ", err)
 	}
-
-	if global.Config.Env.Domain != "" {
-		wh, _ := tg.NewWebhook("https://" + global.Config.Env.Domain + "/wh/tg/" + global.Bot.Token)
-		if _, err = global.Bot.Request(wh); err != nil {
-			global.Log.Errorln("设置webhook失败[oifoghe]: ", err)
-			return
-		}
-
-		info, err := global.Bot.GetWebhookInfo()
-		if err != nil {
-			global.Log.Errorln("获取webhook失败[iuieee]: ", err)
-			return
-		}
-
-		if info.LastErrorDate != 0 {
-			global.Log.Errorln("获取tg信息错误[fosdjfoisj]: ", info.LastErrorMessage)
-			return
-		}
-		global.Log.Printf("成功配置tg[doiasjo]: %s", global.Bot.Self.UserName)
+	if global.Config.Env.Domain == "" {
+		return
 	}
+
+	wh, _ := tg.NewWebhook("https://" + global.Config.Env.Domain + "/wh/tg/" + global.Bot.Token)
+	if _, err = global.Bot.Request(wh); err != nil {
+		global.Log.Errorln("设置webhook失败[oifoghe]: ", err)
+		return
+	}
+
+	info, err := global.Bot.GetWebhookInfo()
+	if err != nil {
+		global.Log.Errorln("获取webhook失败[iuieee]: ", err)
+		return
+	}
+
+	if info.LastErrorDate != 0 {
+		global.Log.Errorln("获取tg信息错误[fosdjfoisj]: ", info.LastErrorMessage)
+		return
+	}
+	global.Log.Printf("成功配置tg[doiasjo]: %s", global.Bot.Self.UserName)
 }
 
 func TgClear() (err error) {
